Separate profile header view file
